Holy crap it works. I ended up removing the additional scaling factor because while it helped in the center it caused issues on the edges.
My erratic readings when rotating was due to a tilt in the sensor -- it pointed about 3-4 degrees behind the vehicle and made it think it was ~8 inches behind where it actually was. This is now compensated for in the code, since correcting that small of an incline for the sensor would be quite difficult.